Circle ci changes

+++ run_test.sh
@@ -2,7 +2,7 @@
 
 # Run Unit Test for mysql
 
-if [[ $MYSQL_TEST == "true" ]]; then
+if [ $MYSQL_TEST == "true" ] && [ $CIRCLECI != "true" ]; then
   # set MYSQL_TEST true if you want to run test against Mysql.
   # you need running mysql on local without root password for testing.
   mysql -uroot -e "drop database if exists gohan_test; create database gohan_test;"
@@ -15,7 +15,7 @@ ETCD_PID=$!
 echo "mode: count" > profile.cov
 
 # Standard go tooling behavior is to ignore dirs with leading underscors
-for dir in $(find . -maxdepth 10 -not -path './.git*' -not -path '*/_*' -not -path './vendor/*' -not -path '*/test_data/*' -not -path '*/goext_example*' -type d);
+for dir in $(find . -maxdepth 10 -not -path './.git*' -not -path '*/_*' -not -path './vendor/*' -not -path '*/test_data/*' -not -path '*/goext_example*' -type d | sort);
 do
 if ls $dir/*.go &> /dev/null; then
     go test -race -covermode=atomic -coverprofile=$dir/profile.tmp $dir
